feat: random rule 방식으로 Game, RacingGame 수정
