+++ src/domains/RacingGame.js
@@ -1,75 +1,47 @@
-import Game from './Game.js'
 import Car from './Car.js'
-import { getRandomNumber } from '../utils/getRandomNumber.js'
-import { racingValidations } from '../validations/racing.js'
+import Game from './Game.js'
 
 export default class RacingGame extends Game {
-  cars = {}
-  carNames = []
-
-  get winners() {
-    const maxPosition = Math.max(
-      ...this.carNames.map(name => this.cars[name].position)
-    )
-    return this.carNames.filter(
-      name => this.cars[name].position === maxPosition
-    )
+  constructor({ carNames, config }) {
+    super(config)
+    this.cars = this.createCars(carNames)
   }
 
-  async setup() {
-    this.initializeCars()
-    const carNames = await this.readCarNames()
-    carNames.forEach(name => this.registerCar(name))
-    this.validate()
+  get winner() {
+    const lastCapturedCars = this.lastResult.cars
+    const maxPosition = Math.max(...Object.values(lastCapturedCars))
+    return Object.keys(lastCapturedCars).filter(name => lastCapturedCars[name] === maxPosition)
   }
 
-  eachRound() {
-    this.display.print('')
-    this.carNames.forEach(name => this.moveCarByRandomNumber(name))
-    this.carNames.forEach(name => this.showCarPosition(name))
+  createCars(carNames) {
+    return carNames.map(name => new Car(name))
   }
 
-  finish() {
-    this.showWinners()
+  doRound() {
+    const ruleName = this.randomRule
+    this.cars.forEach(car => {
+      this.race(ruleName, car)
+    })
+    this.updateResult(ruleName)
   }
 
-  initializeCars() {
-    this.cars = {}
-    this.carNames = []
+  race(ruleName, car) {
+    const ruleResult = this.rules[ruleName]()
+    if (typeof ruleResult === 'number') {
+      car.move(ruleResult)
+    } else if (ruleResult) {
+      car.move(1)
+    }
   }
 
-  async readCarNames() {
-    return (
-      await this.display.read(
-        '경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).'
-      )
-    ).split(',')
-  }
-
-  registerCar(name) {
-    const registeredCar = new Car(name)
-    this.cars[registeredCar.name] = registeredCar
-    this.carNames.push(registeredCar.name)
-  }
-
-  validate() {
-    Object.values(racingValidations).forEach(({ check, errorMessage }) => {
-      if (!check(this.carNames)) {
-        throw new Error(errorMessage)
-      }
+  updateResult(ruleName) {
+    const captureCars = this.cars.reduce((acc, car) => ({
+      ...acc,
+      [car.name]: car.position,
+    }), {})
+    this.addResult({
+      ruleName, cars: captureCars,
     })
   }
 
-  moveCarByRandomNumber(name) {
-    const { min, max, threshold } = this.config
-    if (getRandomNumber(min, max) > threshold) this.cars[name].move()
-  }
-
-  showCarPosition(name) {
-    this.display.print(`${name} : ${'_'.repeat(this.cars[name].position)}`)
-  }
-
-  showWinners() {
-    this.display.print(`\n🎉 우승자 : ${this.winners.join(', ')} 🎉\n`)
-  }
 }
