+++ src/main.js
@@ -1,25 +1,11 @@
-import Printer from './service/Printer.js'
-import NewRacingGame from './domains/NewRacingGame.js'
+import RacingGame from './domains/RacingGame.js'
 import { getRandomNumber } from './utils/getRandomNumber.js'
 
 async function main() {
-  const printer = new Printer()
-
-  // const racingGame = new RacingGame({
-  //   display: printer,
-  //   maxRound: 5,
-  //   config: {
-  //     min: 0,
-  //     max: 9,
-  //     threshold: 4,
-  //   },
-  // })
-  // await racingGame.play()
-
-  const newRacingGame = new NewRacingGame({
+  const newRacingGame = new RacingGame({
     carNames: ['cha', 'cszzi', 'hello'],
     config: {
-      maxRound: 3,
+      maxRound: 5,
       rules: {
         randomNumberMove: () => getRandomNumber(0, 5),
         randomNumberOver40: () => getRandomNumber(0, 100) > 40,
